pyproject.toml
[tool.poetry]
name = "docker-django-redis-celery"
version = "0.1.0"
description = ""
authors = ["VladimirShtefan <shtefanvladimir14@gmail.com>"]
readme = "README.md"
packages = [{include = "docker_django_redis_celery"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
python-dotenv = "^1.0.1"
django-celery-beat = "^2.6.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
